feat: Refactor API integration and enhance user experience
- Updated HardwareKeyModal.vue to use authAPI for WebAuthn credential management.
- Modified ProfileHeader.vue to utilize levelAPI for fetching user level information.
- Enhanced GameRoom.vue to load substance names using substanceAPI.
- Introduced ranked matchmaking and level range settings in Lobby.vue.
- Improved ranking display in Ranking.vue with level badges and win rate information.
- Adjusted Reactions.vue to streamline reaction deletion process.
- Expanded api.ts with new endpoints for level management and substance retrieval.
- Added new WebAuthn password reset and change password functionalities in main.go.

+++ backend/database/models.go
@@ -62,8 +62,8 @@ type User struct {
 	Role               string         `gorm:"default:user;size:20" json:"role"`
 	TwoFactorEnabled   bool           `gorm:"default:false" json:"two_factor_enabled"`
 	TwoFactorSecret    string         `json:"-"`
-	Points             int            `gorm:"default:1000" json:"points"`
-	MonthlyPoints      int            `gorm:"default:1000" json:"monthly_points"`
+	Points             float64        `gorm:"default:1000" json:"points"`
+	MonthlyPoints      float64        `gorm:"default:1000" json:"monthly_points"`
 	Level              int            `gorm:"default:1" json:"level"`
 	XP                 int            `gorm:"default:0" json:"xp"`
 	TotalXP            int            `gorm:"default:0" json:"total_xp"`
